cloud/skeleton/server/requirements.txt
asgiref==3.5.0
Django==3.2.5
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.11.3
Markdown==3.3.6
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
zipp==3.7.0
